Java: Consolidate spring-ldap-2.3.2 stubs.
