--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/core/ContextMapper.java
@@ -1,4 +1,0 @@
-package org.springframework.ldap.core;
-
-public interface ContextMapper<T> {
-}
