--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/core/NameClassPairCallbackHandler.java
@@ -1,3 +1,0 @@
-package org.springframework.ldap.core;
-
-public interface NameClassPairCallbackHandler { }
