--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/filter/EqualsFilter.java
@@ -1,5 +1,0 @@
-package org.springframework.ldap.filter;
-
-public class EqualsFilter implements Filter {
-  public EqualsFilter(String attribute, String value) { }
-}
