--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/filter/HardcodedFilter.java
@@ -1,7 +1,0 @@
-package org.springframework.ldap.filter;
-
-public class HardcodedFilter implements Filter {
-  public HardcodedFilter(String filter) { }
-  public StringBuffer encode(StringBuffer buff) { return buff; }
-  public String toString() { return ""; }
-}
