--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/query/ConditionCriteria.java
@@ -1,5 +1,0 @@
-package org.springframework.ldap.query;
-
-public interface ConditionCriteria {
-  ContainerCriteria is(String value);
-}
