--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/support/LdapEncoder.java
@@ -1,5 +1,0 @@
-package org.springframework.ldap.support;
-
-public class LdapEncoder {
-  public static String filterEncode(String value) { return null; }
-}
