--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/support/LdapNameBuilder.java
@@ -1,12 +1,0 @@
-package org.springframework.ldap.support;
-
-import javax.naming.ldap.LdapName;
-
-public class LdapNameBuilder {
-  public static LdapNameBuilder newInstance() { return null; }
-  public static LdapNameBuilder newInstance(String name) { return null; }
-
-  public LdapNameBuilder add(String name) { return null; }
-  public LdapNameBuilder add(String key, Object value) { return null; }
-  public LdapName build() { return null; }
-}
