--- java/ql/test/experimental/stubs/spring-ldap-2.3.2/org/springframework/ldap/support/LdapUtils.java
@@ -1,7 +1,0 @@
-package org.springframework.ldap.support;
-
-import javax.naming.ldap.LdapName;
-
-public class LdapUtils {
-  public static LdapName newLdapName(String distinguishedName) { return null; }
-}
